[memory][refactor] Relocate sc-base directory
