+++ sc-memory/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(sc-base)
 add_subdirectory(sc-core)
 add_subdirectory(sc-memory)
