+++ sc-memory/sc-core/sc-store/sc-base/sc_message.h
