+++ sc-memory/sc-core/sc-store/sc_iterator5.c
@@ -7,8 +7,8 @@
 #include "sc_iterator.h"
 #include "sc_storage.h"
 
-#include "sc_allocator.h"
-#include "sc_assert_utils.h"
+#include "sc-base/sc_allocator.h"
+#include "sc-base/sc_assert_utils.h"
 
 sc_iterator5 * sc_iterator5_new(
     const sc_memory_context * ctx,
