+++ sc-memory/sc-core/sc-store/sc_segment.c
@@ -9,9 +9,9 @@
 #include "sc_element.h"
 #include "../sc_memory_private.h"
 
-#include "sc_allocator.h"
-#include "sc_atomic.h"
-#include "sc_assert_utils.h"
+#include "sc-base/sc_allocator.h"
+#include "sc-base/sc_atomic.h"
+#include "sc-base/sc_assert_utils.h"
 
 sc_segment * sc_segment_new(sc_addr_seg num)
 {
