+++ sc-memory/sc-core/sc-store/sc_stream_memory.c
@@ -7,8 +7,8 @@
 #include "sc_stream_memory.h"
 #include "sc_stream_private.h"
 
-#include "sc_allocator.h"
-#include "sc_assert_utils.h"
+#include "sc-base/sc_allocator.h"
+#include "sc-base/sc_assert_utils.h"
 
 struct _sc_memory_buffer
 {
