+++ sc-memory/sc-core/sc_helper.c
@@ -9,9 +9,9 @@
 
 #include "string.h"
 
-#include "sc_allocator.h"
-#include "sc_assert_utils.h"
-#include "sc_message.h"
+#include "sc-store/sc-base/sc_allocator.h"
+#include "sc-store/sc-base/sc_assert_utils.h"
+#include "sc-store/sc-base/sc_message.h"
 
 // sc-helper initialization flag
 sc_bool sc_helper_is_initialized = SC_FALSE;
